// packages/database/convex/shared/generated_github_client.ts
export class SimpleUser extends S.Class<SimpleUser>("SimpleUser")({
	name: S.optionalWith(S.String, { nullable: true }),
	email: S.optionalWith(S.String, { nullable: true }),
	login: S.String,
	id: S.Int,
	node_id: S.String,
	avatar_url: S.String,
	gravatar_id: S.NullOr(S.String),
	url: S.String,
	html_url: S.String,
	followers_url: S.String,
	following_url: S.String,
	gists_url: S.String,
	starred_url: S.String,
	subscriptions_url: S.String,
	organizations_url: S.String,
	repos_url: S.String,
	events_url: S.String,
	received_events_url: S.String,
	type: S.String,
	site_admin: S.Boolean,
	starred_at: S.optionalWith(S.String, { nullable: true }),
	user_view_type: S.optionalWith(S.String, { nullable: true }),
}) {}
// ...
export class NullableSimpleUser extends S.Class<NullableSimpleUser>(
	"NullableSimpleUser",
)({
	name: S.optionalWith(S.String, { nullable: true }),
	email: S.optionalWith(S.String, { nullable: true }),
	login: S.String,
	id: S.Int,
	node_id: S.String,
	avatar_url: S.String,
	gravatar_id: S.NullOr(S.String),
	url: S.String,
	html_url: S.String,
	followers_url: S.String,
	following_url: S.String,
	gists_url: S.String,
	starred_url: S.String,
	subscriptions_url: S.String,
	organizations_url: S.String,
	repos_url: S.String,
	events_url: S.String,
	received_events_url: S.String,
	type: S.String,
	site_admin: S.Boolean,
	starred_at: S.optionalWith(S.String, { nullable: true }),
	user_view_type: S.optionalWith(S.String, { nullable: true }),
}) {}
// ...
export class NullableMilestone extends S.Class<NullableMilestone>(
	"NullableMilestone",
)({
	url: S.String,
	html_url: S.String,
	labels_url: S.String,
	id: S.Int,
	node_id: S.String,
	/**
	 * The number of the milestone.
	 */
	number: S.Int,
	/**
	 * The state of the milestone.
	 */
	state: NullableMilestoneState.pipe(
		S.propertySignature,
		S.withConstructorDefault(() => "open" as const),
	),
	/**
	 * The title of the milestone.
	 */
	title: S.String,
	description: S.NullOr(S.String),
	creator: S.NullOr(NullableSimpleUser),
	open_issues: S.Int,
	closed_issues: S.Int,
	created_at: S.String,
	updated_at: S.String,
	closed_at: S.NullOr(S.String),
	due_on: S.NullOr(S.String),
}) {}

/**
 * Hypermedia Link
 */
export class Link extends S.Class<Link>("Link")({
	href: S.String,
}) {}

/**
 * How the author is associated with the repository.
 */
export class AuthorAssociation extends S.Literal(
	"COLLABORATOR",
	"CONTRIBUTOR",
	"FIRST_TIMER",
	"FIRST_TIME_CONTRIBUTOR",
	"MANNEQUIN",
	"MEMBER",
	"NONE",
	"OWNER",
) {}
// ...
export class PullRequestSimple extends S.Class<PullRequestSimple>(
	"PullRequestSimple",
)({
	url: S.String,
	id: S.Int,
	node_id: S.String,
	html_url: S.String,
	diff_url: S.String,
	patch_url: S.String,
	issue_url: S.String,
	commits_url: S.String,
	review_comments_url: S.String,
	review_comment_url: S.String,
	comments_url: S.String,
	statuses_url: S.String,
	number: S.Int,
	state: S.String,
	locked: S.Boolean,
	title: S.String,
	user: S.NullOr(NullableSimpleUser),
	body: S.NullOr(S.String),
	labels: S.Array(
		S.Struct({
			id: S.Int,
			node_id: S.String,
			url: S.String,
			name: S.String,
			description: S.NullOr(S.String),
			color: S.String,
			default: S.Boolean,
		}),
	),
	milestone: S.NullOr(NullableMilestone),
	active_lock_reason: S.optionalWith(S.String, { nullable: true }),
	created_at: S.String,
	updated_at: S.String,
	closed_at: S.NullOr(S.String),
	merged_at: S.NullOr(S.String),
	merge_commit_sha: S.NullOr(S.String),
	assignee: S.NullOr(NullableSimpleUser),
	assignees: S.optionalWith(S.Array(SimpleUser), { nullable: true }),
	requested_reviewers: S.optionalWith(S.Array(SimpleUser), { nullable: true }),
	requested_teams: S.optionalWith(S.Array(Team), { nullable: true }),
	head: S.Struct({
		label: S.String,
		ref: S.String,
		repo: S.NullOr(Repository),
		sha: S.String,
		user: S.NullOr(NullableSimpleUser),
	}),
	base: S.Struct({
		label: S.String,
		ref: S.String,
		repo: S.NullOr(Repository),
		sha: S.String,
		user: S.NullOr(NullableSimpleUser),
	}),
	_links: S.Struct({
		comments: Link,
		commits: Link,
		statuses: Link,
		html: Link,
		issue: Link,
		review_comments: Link,
		review_comment: Link,
		self: Link,
	}),
	author_association: AuthorAssociation,
	auto_merge: S.NullOr(AutoMerge),
	/**
	 * Indicates whether or not the pull request is a draft.
	 */
	draft: S.optionalWith(S.Boolean, { nullable: true }),
}) {}
// ...
export class ReactionRollup extends S.Class<ReactionRollup>("ReactionRollup")({
	url: S.String,
	total_count: S.Int,
	"+1": S.Int,
	"-1": S.Int,
	laugh: S.Int,
	confused: S.Int,
	heart: S.Int,
	hooray: S.Int,
	eyes: S.Int,
	rocket: S.Int,
}) {}
// ...
export class IssueStateReason extends S.Literal(
	"completed",
	"reopened",
	"not_planned",
	"duplicate",
) {}

/**
 * The color of the issue type.
 */
export class IssueTypeColor extends S.Literal(
	"gray",
	"blue",
	"green",
	"yellow",
	"orange",
	"red",
	"pink",
	"purple",
) {}

/**
 * The type of issue.
 */
export class IssueType extends S.Class<IssueType>("IssueType")({
	/**
	 * The unique identifier of the issue type.
	 */
	id: S.Int,
	/**
	 * The node identifier of the issue type.
	 */
	node_id: S.String,
	/**
	 * The name of the issue type.
	 */
	name: S.String,
	/**
	 * The description of the issue type.
	 */
	description: S.NullOr(S.String),
	/**
	 * The color of the issue type.
	 */
	color: S.optionalWith(IssueTypeColor, { nullable: true }),
	/**
	 * The time the issue type created.
	 */
	created_at: S.optionalWith(S.String, { nullable: true }),
	/**
	 * The time the issue type last updated.
	 */
	updated_at: S.optionalWith(S.String, { nullable: true }),
	/**
	 * The enabled state of the issue type.
	 */
	is_enabled: S.optionalWith(S.Boolean, { nullable: true }),
}) {}

/**
 * GitHub apps are a new way to extend GitHub. They can be installed directly on organizations and user accounts and granted access to specific repositories. They come with granular permissions and built-in webhooks. GitHub apps are first class actors within GitHub.
 */
export class NullableIntegration extends S.Class<NullableIntegration>(
	"NullableIntegration",
)({
	/**
	 * Unique identifier of the GitHub app
	 */
	id: S.Int,
	/**
	 * The slug name of the GitHub app
	 */
	slug: S.optionalWith(S.String, { nullable: true }),
	node_id: S.String,
	client_id: S.optionalWith(S.String, { nullable: true }),
	owner: S.Union(SimpleUser, Enterprise),
	/**
	 * The name of the GitHub app
	 */
	name: S.String,
	description: S.NullOr(S.String),
	external_url: S.String,
	html_url: S.String,
	created_at: S.String,
	updated_at: S.String,
	/**
	 * The set of permissions for the GitHub app
	 */
	permissions: S.Struct({
		issues: S.optionalWith(S.String, { nullable: true }),
		checks: S.optionalWith(S.String, { nullable: true }),
		metadata: S.optionalWith(S.String, { nullable: true }),
		contents: S.optionalWith(S.String, { nullable: true }),
		deployments: S.optionalWith(S.String, { nullable: true }),
	}),
	/**
	 * The list of events for the GitHub app. Note that the `installation_target`, `security_advisory`, and `meta` events are not included because they are global events and not specific to an installation.
	 */
	events: S.Array(S.String),
	/**
	 * The number of installations associated with the GitHub app. Only returned when the integration is requesting details about itself.
	 */
	installations_count: S.optionalWith(S.Int, { nullable: true }),
}) {}

export class SubIssuesSummary extends S.Class<SubIssuesSummary>(
	"SubIssuesSummary",
)({
	total: S.Int,
	completed: S.Int,
	percent_completed: S.Int,
}) {}

/**
 * Context around who pinned an issue comment and when it was pinned.
 */
export class NullablePinnedIssueComment extends S.Class<NullablePinnedIssueComment>(
	"NullablePinnedIssueComment",
)({
	pinned_at: S.String,
	pinned_by: S.NullOr(NullableSimpleUser),
}) {}

/**
 * Comments provide a way for people to collaborate on an issue.
 */
export class NullableIssueComment extends S.Class<NullableIssueComment>(
	"NullableIssueComment",
)({
	/**
	 * Unique identifier of the issue comment
	 */
	id: S.Int,
	node_id: S.String,
	/**
	 * URL for the issue comment
	 */
	url: S.String,
	/**
	 * Contents of the issue comment
	 */
	body: S.optionalWith(S.String, { nullable: true }),
	body_text: S.optionalWith(S.String, { nullable: true }),
	body_html: S.optionalWith(S.String, { nullable: true }),
	html_url: S.String,
	user: S.NullOr(NullableSimpleUser),
	created_at: S.String,
	updated_at: S.String,
	issue_url: S.String,
	author_association: S.optionalWith(AuthorAssociation, { nullable: true }),
	performed_via_github_app: S.optionalWith(NullableIntegration, {
		nullable: true,
	}),
	reactions: S.optionalWith(ReactionRollup, { nullable: true }),
	pin: S.optionalWith(NullablePinnedIssueComment, { nullable: true }),
}) {}

export class IssueDependenciesSummary extends S.Class<IssueDependenciesSummary>(
	"IssueDependenciesSummary",
)({
	blocked_by: S.Int,
	blocking: S.Int,
	total_blocked_by: S.Int,
	total_blocking: S.Int,
}) {}

/**
 * The data type of the issue field
 */
export class IssueFieldValueDataType extends S.Literal(
	"text",
	"single_select",
	"number",
	"date",
) {}

/**
 * A value assigned to an issue field
 */
export class IssueFieldValue extends S.Class<IssueFieldValue>(
	"IssueFieldValue",
)({
	/**
	 * Unique identifier for the issue field.
	 */
	issue_field_id: S.Int,
	node_id: S.String,
	/**
	 * The data type of the issue field
	 */
	data_type: IssueFieldValueDataType,
	/**
	 * The value of the issue field
	 */
	value: S.NullOr(S.Union(S.String, S.Number, S.Int)),
	/**
	 * Details about the selected option (only present for single_select fields)
	 */
	single_select_option: S.optionalWith(
		S.Struct({
			/**
			 * Unique identifier for the option.
			 */
			id: S.Int,
			/**
			 * The name of the option
			 */
			name: S.String,
			/**
			 * The color of the option
			 */
			color: S.String,
		}),
		{ nullable: true },
	),
}) {}

/**
 * Issues are a great way to keep track of tasks, enhancements, and bugs for your projects.
 */
export class Issue extends S.Class<Issue>("Issue")({
	id: S.Int,
	node_id: S.String,
	/**
	 * URL for the issue
	 */
	url: S.String,
	repository_url: S.String,
	labels_url: S.String,
	comments_url: S.String,
	events_url: S.String,
	html_url: S.String,
	/**
	 * Number uniquely identifying the issue within its repository
	 */
	number: S.Int,
	/**
	 * State of the issue; either 'open' or 'closed'
	 */
	state: S.String,
	/**
	 * The reason for the current state
	 */
	state_reason: S.optionalWith(IssueStateReason, { nullable: true }),
	/**
	 * Title of the issue
	 */
	title: S.String,
	/**
	 * Contents of the issue
	 */
	body: S.optionalWith(S.String, { nullable: true }),
	user: S.NullOr(NullableSimpleUser),
	/**
	 * Labels to associate with this issue; pass one or more label names to replace the set of labels on this issue; send an empty array to clear all labels from the issue; note that the labels are silently dropped for users without push access to the repository
	 */
	labels: S.Array(
		S.Union(
			S.String,
			S.Struct({
				id: S.optionalWith(S.Int, { nullable: true }),
				node_id: S.optionalWith(S.String, { nullable: true }),
				url: S.optionalWith(S.String, { nullable: true }),
				name: S.optionalWith(S.String, { nullable: true }),
				description: S.optionalWith(S.String, { nullable: true }),
				color: S.optionalWith(S.String, { nullable: true }),
				default: S.optionalWith(S.Boolean, { nullable: true }),
			}),
		),
	),
	assignee: S.NullOr(NullableSimpleUser),
	assignees: S.optionalWith(S.Array(SimpleUser), { nullable: true }),
	milestone: S.NullOr(NullableMilestone),
	locked: S.Boolean,
	active_lock_reason: S.optionalWith(S.String, { nullable: true }),
	comments: S.Int,
	pull_request: S.optionalWith(
		S.Struct({
			merged_at: S.optionalWith(S.String, { nullable: true }),
			diff_url: S.NullOr(S.String),
			html_url: S.NullOr(S.String),
			patch_url: S.NullOr(S.String),
			url: S.NullOr(S.String),
		}),
		{ nullable: true },
	),
	closed_at: S.NullOr(S.String),
	created_at: S.String,
	updated_at: S.String,
	draft: S.optionalWith(S.Boolean, { nullable: true }),
	closed_by: S.optionalWith(NullableSimpleUser, { nullable: true }),
	body_html: S.optionalWith(S.String, { nullable: true }),
	body_text: S.optionalWith(S.String, { nullable: true }),
	timeline_url: S.optionalWith(S.String, { nullable: true }),
	type: S.optionalWith(IssueType, { nullable: true }),
	repository: S.optionalWith(Repository, { nullable: true }),
	performed_via_github_app: S.optionalWith(NullableIntegration, {
		nullable: true,
	}),
	author_association: S.optionalWith(AuthorAssociation, { nullable: true }),
	reactions: S.optionalWith(ReactionRollup, { nullable: true }),
	sub_issues_summary: S.optionalWith(SubIssuesSummary, { nullable: true }),
	/**
	 * URL to get the parent issue of this issue, if it is a sub-issue
	 */
	parent_issue_url: S.optionalWith(S.String, { nullable: true }),
	pinned_comment: S.optionalWith(NullableIssueComment, { nullable: true }),
	issue_dependencies_summary: S.optionalWith(IssueDependenciesSummary, {
		nullable: true,
	}),
	issue_field_values: S.optionalWith(S.Array(IssueFieldValue), {
		nullable: true,
	}),
}) {}